\name{Wret}
\alias{Wret}
\docType{data}
\title{Water retention data from Awach watershed in the lake Victoria basin, Western Kenya}
\description{
  This is part of the water retention characteristics from a study 
of soil physical properties in the Lake Victoria basin in Western Kenya  }
\usage{data(Wret)}
\format{
  A data frame with 232 observations on the following 6 variables
  \describe{
    \item{\code{Points}}{an ordered factor of plotcodes  }
    \item{\code{Depth}}{a factor with levels \code{Subsoil} \code{Topsoil}  }
    \item{\code{BULK}}{a numeric vector of bulk density in g/cm3  }
    \item{\code{x}}{a numeric vector of suction potential in cm  }
    \item{\code{y}}{a numeric vector of volumetric moisture content in vol/vol  }
    \item{\code{lnx}}{a numeric vector of logarithm of suction potential  }
  }
}
\source{
  Omuto, C.T. 2003
 \emph{ Rapid Mapping of saturated hydraulic conductivity in tropical watersheds  }
 \bold{ MSc Thesis}
 Nairobi: University of Nairobi
}
\keyword{datasets}
